pyproject.toml
[project]
name = "gps-imu-sim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.9.2",
    "numpy>=2.1.3",
    "paho-mqtt>=2.1.0",
]
